I cannot concisely explain this edge case but it's fixed
There was a problem in EditBuffer_Cursor_wrangle where I accidentally swapped
row and column, this prevented the function from properly bounds checking. This
led to an edge case where if one cursor deleted the line another was on, it
would segfault when the second cursor attempted to get its current line while
bounds checking itself. This has been fixed.