Fixed out of bounds array element set in String_addRune
String_addRune would append a rune first and then re-allocate the buffer. This
was fine when String used a null terminated c-string with at least one byte
guaranteed to exist at the end of the buffer, but as that currently no longer
exists it needed to be fixed. This causes valgrind to complain *considerably*
less, and the text actually renders properly when using it.