Interface and tab modules now implement new recalculate/redraw model
Doing a bit of work to overhaul the interface rendering system to make it more
efficient and easier to work with. Operations that mutate interface elements
will set recalculate or redraw flags, and at the end of an event listener, these
will be fulfilled by recursively "refreshing" all elements. Consolidating
rendering operations in one place will also help, in the future, to hold off on
rendering anything until there are no events in queue.